sql-injection/requirements.txt
Flask
mysql_connector_repackaged
peewee
psycopg2_binary
pydantic
pytest
SQLAlchemy
